requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
web3==6.9.0
pymongo==4.5.0
pydantic==2.3.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
email-validator==2.0.0.post2
python-dotenv==1.0.0
streamlit==1.26.0
qrcode==7.4.2
pandas==2.1.0
pytest==7.4.2
httpx==0.25.0